unit tests pass

+++ src/warehouses/snowflake_warehouse.py
@@ -254,7 +254,10 @@ class SnowflakeWarehouse(AbstractWarehouse):
             "records_to_insert": None
         }
 
+        print(f"{table_info["table"]}: 1")
+
         if cdc_type in ("APPEND_ONLY_STREAM", "STANDARD_STREAM"):
+            print(f"{table_info["table"]}: 2")
 
             self._cleanup_records(table_info, existing_etl_ids)
 
@@ -262,10 +265,12 @@ class SnowflakeWarehouse(AbstractWarehouse):
             stream_name = self.get_stream_name(table_info)
             
             # Load stream data into processing table
-            self.cursor.execute(f"INSERT INTO {stream_processing_table_name} SELECT * FROM {stream_name};")
-            self.cursor.execute(f"UPDATE TABLE {stream_processing_table_name} SET etl_id = '{new_etl_id}'")
+            self.cursor.execute(f"INSERT INTO {stream_processing_table_name} SELECT *, '{new_etl_id}' FROM {stream_name};")
+            self.cursor.execute(f"UPDATE {stream_processing_table_name} SET etl_id = '{new_etl_id}'")
 
             if cdc_type == "STANDARD_STREAM":
+                print(f"{table_info["table"]}: 3")
+
                 # For standard streams, get primary keys of records to delete
                 primary_keys = self.get_primary_keys(table_info)
                 if primary_keys == []:
@@ -278,6 +283,9 @@ class SnowflakeWarehouse(AbstractWarehouse):
                 """
                 updates_dict['records_to_delete'] = self.get_df_batches(delete_query)
 
+            print(f"{table_info["table"]}: 4")
+
+
             # Get records to insert (for both stream types)
             insert_query = f"""
                 SELECT * 
@@ -286,6 +294,8 @@ class SnowflakeWarehouse(AbstractWarehouse):
             """
             raw_batches = self.get_df_batches(insert_query)
             processed_batches = []
+
+            print(5)
             for batch in raw_batches:
                 batch.rename(columns={
                     "METADATA$ROW_ID": "MELCHI_ROW_ID",
@@ -295,6 +305,7 @@ class SnowflakeWarehouse(AbstractWarehouse):
             updates_dict['records_to_insert'] = processed_batches
 
         elif cdc_type == "FULL_REFRESH":
+            print(6)
             # For full refresh, we only need insert records
             updates = self.get_df_batches(
                 f"SELECT * FROM {self.get_full_table_name(table_info)}"
@@ -303,9 +314,12 @@ class SnowflakeWarehouse(AbstractWarehouse):
             for df in updates:
                 all_updates.append(df)
             updates_dict['records_to_insert'] = all_updates
+        print(7)
         return updates_dict
 
     def _cleanup_records(self, table_info, existing_etl_ids):
+        if len(existing_etl_ids) == 0:
+            return
         formatted_ids = [f"'{id}'" for id in existing_etl_ids]
         formatted_where_clause = f"WHERE etl_id in ({', '.join(formatted_ids)})"
         delete_transferred_records_query = f"DELETE FROM {self.get_stream_processing_table_name(table_info)} {formatted_where_clause}"
